Refactor ReplicationManager and Introduce Backup Cluster interfaces.
