+++ ambry-replication/src/main/java/com.github.ambry.replication/ReplicaTokenPersistor.java
@@ -1,0 +1,30 @@
+/**
+ * Copyright 2016 LinkedIn Corp. All rights reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ */
+package com.github.ambry.replication;
+
+import java.io.IOException;
+
+
+/**
+ * {@link ReplicaTokenPersistor} is used in {@link ReplicationEngine} to persist replication token.
+ */
+interface ReplicaTokenPersistor extends Runnable {
+
+  void write(String mountPath, boolean shuttingDown) throws IOException, ReplicationException;
+
+  void write(boolean shuttingDown) throws IOException, ReplicationException;
+
+  void run();
+}
+
