+++ build.gradle
@@ -219,7 +219,8 @@ project(':ambry-replication') {
                 project(':ambry-store'),
                 project(':ambry-commons'),
                 project(':ambry-protocol'),
-                project(':ambry-network')
+                project(':ambry-network'),
+                project(':ambry-cloud')
         compile "com.codahale.metrics:metrics-core:$metricsVersion"
         testCompile project(':ambry-clustermap').sourceSets.test.output
         testCompile project(':ambry-api').sourceSets.test.output
